Added support for interactive authentication.
This PR adds support for interactive authentication flows for the adal4j
library. The browser UI component is implemented using Eclipse SWT's cross
platform browser component. The browser UI is implemented in the
adal-ad-interactive-auth module. All the new code for implementing interactive
auth in the adal4j module is in the com.microsoft.aad.adal4j.interactive.*
package.

The maven project is now a multi-module maven project containing the
browser UI module and the adal4j module. The browser UI module -
adal-ad-interactive-auth - is built and the binaries are hosted in the
github repo for all supported platforms and architectures (Linux, OSX and
Windows for x86 and x64 architectures). At runtime, adal4j picks the
appropriate binary for the current platform and downloads and caches the
jar (uses HTTP ETags to automatically update the local cached version of
the jar in case a new version is found on the server). While the current
host for these jars is Github we can move this to an alternate location if
necessary.