Added UserInfo.tenantID to equals/hashCode comparison.
A new variable, tenantID, was added to the UserInfo class. This variable was not added to the equals/hashCode method to test against so two instance of UserInfo that had different tenantIDs could produce an equals result in error. Also added a test case in an attempt to future proof additional new variables to the UserInfo class.