Travis CI Failure - Parsing uses a calendar instance
When 2 objects are created in succession they might have slightly different values for new GregorianCalendar() because a new instance is created and they get the system time in miliseconds. This could cause two UserInfos for the test to be created at the same time to the user but at different times for the system.